Add warning to Readme
We want to make it completely clear why this fork has been made,
as well as its intended lifespan. I intend to pull in the same
changes which were merged to the master branch, so there shouldn't
be any code in here which doesn't eventually appear in the main fork.

I'm hoping that these should be sufficient for our requirements,
however as there are additional changes on master it may not be
quite as straight forward as I'm hoping.